+++ src/main/lib/contentmanager_fixed.h
@@ -29,7 +29,7 @@ public:
     ContentManager_Fixed(std::unique_ptr<Protocol> _protocol, CommonHeaders &_headerHandler, bool isServer);
     virtual ~ContentManager_Fixed();
     virtual ContentManagerType getType() const noexcept override;
-    void setMaximumSize(unsigned size) noexcept override;
+    void setMaximumSize(unsigned long size) noexcept override;
 protected:
     virtual std::vector<char> ProcessContent(const std::vector<char> & incomingData) noexcept override;
     virtual bool PrepareContent() noexcept override;
